core&ui: move the function to a new file to avoid front build error
