+++ packages/hydrooj/src/handler/contest.ts
@@ -4,10 +4,8 @@ import { stringify as toCSV } from 'csv-stringify/sync';
 import { escapeRegExp, pick } from 'lodash';
 import moment from 'moment-timezone';
 import { ObjectId } from 'mongodb';
-import {
-    Counter, diffArray, randomstring, sortFiles, Time, yaml,
-    getContestProblemAlphabeticId
-} from '@hydrooj/utils/lib/utils';
+import { Counter, diffArray, randomstring, sortFiles, Time, yaml } from '@hydrooj/utils/lib/utils';
+import { getContestProblemAlphabeticId } from '@hydrooj/utils/lib/broswer-can-use';
 import { Context, Service } from '../context';
 import {
     BadRequestError, ContestNotAttendedError, ContestNotEndedError, ContestNotFoundError, ContestNotLiveError,
