+++ packages/utils/lib/utils.ts
@@ -397,17 +397,4 @@ export async function pipeRequest(req: superagent.Request, w: fs.WriteStream, ti
     }
 }
 
-export function getContestProblemAlphabeticId(index: number) {
-  // A...Z, AA...AZ, BA...BZ, ...
-  if (index < 0) return '?';
-  let letters = '';
-  index ++;
-  while (index > 0) {
-    index--;
-    letters = String.fromCharCode(65 + (index % 26)) + letters;
-    index = Math.floor(index / 26);
-  }
-  return letters;
-}
-
 export * from '@hydrooj/utils/lib/common';
